Refactor the server part of the backend out
There's a lot going on in backend.ex. This moves the server piece out so
that the behaviour can stand on its own.
